# specs/openapi/bundled.yaml
openapi: "3.1.0"
info:
  title: Consensus Agent Runtime
  description: |
    Database-native cognitive architecture for AI agents.

    All state lives in the database. The API writes to the same tables the agent reads from.
    There is no separate API state — the database is the single source of truth.

    ## Key Concepts
    - **Sessions** — Every agent instance is a session row. RLS isolation, billing, lifecycle.
    - **Memory Events** — Immutable append-only ledger. Display state managed via display_modes.
    - **Transactions** — Atomic cognition: every thought is a SQL transaction.
    - **Tools & Skills** — Progressive disclosure: metadata always visible, full instructions on demand.
    - **HITL Approvals** — Human-in-the-loop: pause for review before destructive actions.

    ## Authentication
    All endpoints except `/api/v1/health` require a Bearer API key:
    - cs_ak_* : admin — full CRUD
    - cs_sk_* : session — scoped to one session
    - cs_rk_* : readonly — SELECT only
    - cs_wk_* : webhook — event ingestion only

    Keys are SHA-256 hashed before storage. The full key is only returned at creation time.
  version: "0.1.0"
  contact:
    name: Consensus
    url: https://github.com/wojons/consensus
servers:
  - url: http://localhost:8090
    description: Local development (SQLite or local Postgres)
  - url: https://{host}
    description: Remote deployment (any Postgres provider)
    variables:
      host:
        default: localhost:8090
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/sessions:
    post:
      tags: [Sessions]
      summary: Create a new agent session
      operationId: createSession
      security: [{BearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Sessions]
      summary: List sessions
      operationId: listSessions
      security: [{BearerAuth: []}]
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: Filter by comma-separated status values.
      responses:
        "200":
          description: List of sessions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{sessionId}:
    get:
      tags: [Sessions]
      summary: Get session details
      operationId: getSession
      security: [{BearerAuth: []}]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Sessions]
      summary: Update session (pause/resume/cancel)
      operationId: updateSession
      security: [{BearerAuth: []}]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionRequest"
      responses:
        "200":
          description: Updated session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Sessions]
      summary: Soft-delete a session (admin)
      operationId: deleteSession
      security: [{BearerAuth: []}]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session deleted.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{sessionId}/message:
    post:
      tags: [Sessions]
      summary: Send a message to an agent session
      operationId: sendMessage
      security: [{BearerAuth: []}]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Message received.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageReceivedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Memory & Context (SPEC-015 §3.2) ──
  /api/v1/sessions/{sessionId}/memory:
    get:
      tags: [Memory]
      summary: List memory events
      operationId: listMemory
      security: [{BearerAuth: []}]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            type: string
          description: Filter by event type.
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: List of memory events.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryEvent"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{sessionId}/memory/{memoryId}:
    get:
      tags: [Memory]
      summary: Get a single memory event
      operationId: getMemoryEvent
      security: [{BearerAuth: []}]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memoryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Memory event.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEvent"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{sessionId}/context:
    get:
      tags: [Memory]
      summary: Get active context (rendered)
      operationId: getActiveContext
      security: [{BearerAuth: []}]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 200
      responses:
        "200":
          description: Rendered active context items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActiveContextItem"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{sessionId}/iterations:
    get:
      tags: [Memory]
      summary: List iteration commits
      operationId: listIterations
      security: [{BearerAuth: []}]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: List of iteration commits.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IterationCommit"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Tasks (SPEC-015 §3.3) ──
  /api/v1/sessions/{sessionId}/tasks:
    post:
      tags: [Tasks]
      summary: Create a task
      operationId: createTask
      security: [{BearerAuth: []}]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Tasks]
      summary: List tasks for a session
      operationId: listTasks
      security: [{BearerAuth: []}]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status.
      responses:
        "200":
          description: List of tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tasks/{taskId}:
    patch:
      tags: [Tasks]
      summary: Update task status
      operationId: updateTask
      security: [{BearerAuth: []}]
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tasks/{taskId}/claim:
    post:
      tags: [Tasks]
      summary: Claim a task
      operationId: claimTask
      security: [{BearerAuth: []}]
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task claimed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Tools & Skills (SPEC-015 §3.4) ──
  /api/v1/tools:
    get:
      tags: [Tools]
      summary: List available tools
      operationId: listTools
      security: [{BearerAuth: []}]
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: Filter by tool status.
        - name: hemisphere
          in: query
          schema:
            type: string
            enum: [internal, external]
      responses:
        "200":
          description: List of tools.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tool"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tools/{toolName}/execute:
    post:
      tags: [Tools]
      summary: Execute a tool
      operationId: executeTool
      security: [{BearerAuth: []}]
      parameters:
        - name: toolName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteToolRequest"
      responses:
        "200":
          description: Tool execution result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecuteToolResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/skills:
    get:
      tags: [Tools]
      summary: List skill metadata
      operationId: listSkills
      security: [{BearerAuth: []}]
      responses:
        "200":
          description: List of skills (metadata only).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/skills/{skillName}:
    get:
      tags: [Tools]
      summary: Get full skill instructions
      operationId: getSkill
      security: [{BearerAuth: []}]
      parameters:
        - name: skillName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full skill details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── HITL Approvals (SPEC-015 §3.5) ──
  /api/v1/approvals:
    get:
      tags: [Approvals]
      summary: List pending approvals
      operationId: listApprovals
      security: [{BearerAuth: []}]
      parameters:
        - name: session_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of pending approvals.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Approval"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/approvals/{approvalId}:
    get:
      tags: [Approvals]
      summary: Get approval details
      operationId: getApproval
      security: [{BearerAuth: []}]
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Approval details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key prefixed with cs_ (e.g., cs_sk_a1b2c3d4...).
...